+++ messageformat/main.go
@@ -3,10 +3,60 @@ Package messageformat formats messages into a JSON string.
 */
 package messageformat
 
+import (
+	"encoding/json"
+	"fmt"
+	"reflect"
+	"sync"
+)
+
 // ----------------------------------------------------------------------------
-// Interfaces
+// Types
 // ----------------------------------------------------------------------------
 
 type MessageFormatInterface interface {
 	Message(id string, status string, text string, details ...interface{}) (string, error)
+}
+
+// ----------------------------------------------------------------------------
+// Variables
+// ----------------------------------------------------------------------------
+
+var lock sync.Mutex
+
+// ----------------------------------------------------------------------------
+// Internal functions
+// ----------------------------------------------------------------------------
+
+func isJson(unknownString string) bool {
+	var jsonString json.RawMessage
+	return json.Unmarshal([]byte(unknownString), &jsonString) == nil
+}
+
+func jsonAsInterface(unknownString string) interface{} {
+	var jsonString json.RawMessage
+	json.Unmarshal([]byte(unknownString), &jsonString)
+	return jsonString
+}
+
+func stringify(unknown interface{}) string {
+	// See https://pkg.go.dev/fmt for format strings.
+	var result string
+	switch value := unknown.(type) {
+	case string:
+		result = value
+	case int:
+		result = fmt.Sprintf("%d", value)
+	case float64:
+		result = fmt.Sprintf("%g", value)
+	case bool:
+		result = fmt.Sprintf("%t", value)
+	case error:
+		result = value.Error()
+	default:
+		xType := reflect.TypeOf(unknown)
+		xValue := reflect.ValueOf(unknown)
+		result = fmt.Sprintf("(%s)%+v", xType, xValue)
+	}
+	return result
 }
