+++ messagelogger/main.go
@@ -1,5 +1,5 @@
 /*
-Package logger provides...
+Package messagelogger generates and logs messages.
 */
 package messagelogger
 
@@ -17,25 +17,6 @@ import (
 // ----------------------------------------------------------------------------
 
 type Level int
-
-// ----------------------------------------------------------------------------
-// Constants
-// ----------------------------------------------------------------------------
-
-// Must match what's in logger/main.go
-const (
-	LevelTrace Level = iota
-	LevelDebug
-	LevelInfo
-	LevelWarn
-	LevelError
-	LevelFatal
-	LevelPanic
-)
-
-// ----------------------------------------------------------------------------
-// Interfaces
-// ----------------------------------------------------------------------------
 
 type MessageLoggerInterface interface {
 	Error(messageNumber int, details ...interface{}) error
@@ -56,6 +37,21 @@ type MessageLoggerInterface interface {
 }
 
 // ----------------------------------------------------------------------------
+// Constants
+// ----------------------------------------------------------------------------
+
+// Must match what's in logger/main.go
+const (
+	LevelTrace Level = iota
+	LevelDebug
+	LevelInfo
+	LevelWarn
+	LevelError
+	LevelFatal
+	LevelPanic
+)
+
+// ----------------------------------------------------------------------------
 // Variables
 // ----------------------------------------------------------------------------
 
@@ -65,10 +61,9 @@ var messageLoggerInstance *MessageLoggerDefault
 // Constructors
 // ----------------------------------------------------------------------------
 
-// TODO:
 func New() *MessageLoggerDefault {
 	result := &MessageLoggerDefault{
-		Logger:          &logger.LoggerImpl{},
+		Logger:          &logger.LoggerDefault{},
 		MessageFormat:   &messageformat.MessageFormatJson{},
 		MessageId:       &messageid.MessageIdDefault{},
 		MessageLogLevel: &messageloglevel.MessageLogLevelDefault{},
@@ -83,88 +78,82 @@ func New() *MessageLoggerDefault {
 // Internal functions
 // ----------------------------------------------------------------------------
 
-// TODO:
 func init() {
 	messageLoggerInstance = New()
 }
 
 // ----------------------------------------------------------------------------
-// Public Functions
+// Public functions
 // ----------------------------------------------------------------------------
 
-// TODO:
 func Error(messageNumber int, details ...interface{}) error {
 	return messageLoggerInstance.Error(messageNumber, details...)
 }
 
-// TODO:
-func GetLogLevel() Level { return messageLoggerInstance.GetLogLevel() }
+func GetLogLevel() Level {
+	return messageLoggerInstance.GetLogLevel()
+}
 
-// TODO:
-func GetLogLevelAsString() string { return messageLoggerInstance.GetLogLevelAsString() }
+func GetLogLevelAsString() string {
+	return messageLoggerInstance.GetLogLevelAsString()
+}
 
-// TODO:
 func Log(messageNumber int, details ...interface{}) error {
 	return messageLoggerInstance.Log(messageNumber, details...)
 }
 
-// TODO:
 func Message(messageNumber int, details ...interface{}) (string, error) {
 	return messageLoggerInstance.Message(messageNumber, details...)
 }
 
-// TODO:
-func GetMessageLogger() *MessageLoggerDefault { return messageLoggerInstance }
+func MsgNumber(messageNumber int) messagetext.MessageNumber {
+	return messagetext.MsgNumber(messageNumber)
+}
 
-// TODO:
+func GetMessageLogger() *MessageLoggerDefault {
+	return messageLoggerInstance
+}
+
 func SetIdTemplate(idTemplate string) MessageLoggerInterface {
 	return messageLoggerInstance.SetIdTemplate(idTemplate)
 }
 
-// TODO:
 func SetLogger(logger logger.LoggerInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetLogger(logger)
 }
 
-// TODO:
-func SetLogLevel(level Level) MessageLoggerInterface { return messageLoggerInstance.SetLogLevel(level) }
+func SetLogLevel(level Level) MessageLoggerInterface {
+	return messageLoggerInstance.SetLogLevel(level)
+}
 
-// TODO:
 func SetLogLevelFromString(levelString string) MessageLoggerInterface {
 	return messageLoggerInstance.SetLogLevelFromString(levelString)
 }
 
-// TODO:
 func SetMessageFormat(messageFormat messageformat.MessageFormatInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetMessageFormat(messageFormat)
 }
 
-// TODO:
 func SetMessageId(messageId messageid.MessageIdInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetMessageId(messageId)
 }
 
-// TODO:
 func SetMessageLogger(messageLogger *MessageLoggerDefault) {
 	messageLoggerInstance = messageLogger
 }
 
-// TODO:
 func SetMessageLogLevel(messageLogLevel messageloglevel.MessageLogLevelInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetMessageLogLevel(messageLogLevel)
 }
 
-// TODO:
 func SetMessageStatus(messageStatus messagestatus.MessageStatusInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetMessageStatus(messageStatus)
 }
 
-// TODO:
 func SetMessageText(messageText messagetext.MessageTextInterface) MessageLoggerInterface {
 	return messageLoggerInstance.SetMessageText(messageText)
 }
 
-// TODO:
 func SetTextTemplates(messages map[int]string) MessageLoggerInterface {
 	return messageLoggerInstance.SetTextTemplates(messages)
 }
