+++ messagestatus/messagestatus_null.go
@@ -1,5 +1,5 @@
 /*
-MessageStatusNull returns an empty string for a status value.
+The MessageStatusNull implementation returns an empty string for a status value.
 */
 package messagestatus
 
@@ -13,9 +13,7 @@ type MessageStatusNull struct{}
 // Interface methods
 // ----------------------------------------------------------------------------
 
-/*
-Get the "status" value given the message id and it's details.
-*/
+// Get the "status" value given the message id and it's details.
 func (messagelevel *MessageStatusNull) MessageStatus(messageNumber int, details ...interface{}) (string, error) {
 	var err error = nil
 	var result = ""
